Factor out trace handling into a module
